Moved user interfaces to a user directory
