+++ src/interfaces/apollo/users/IAuthenticatedUser.ts
