+++ src/interfaces/apollo/users/IUserCreateInput.ts
@@ -1,8 +1,7 @@
 interface IUserCreateInput {
   email: string
   password: string
-  type: string
-  websiteIds: string[]
+  type: 'admin' | 'user'
 }
 
 export default IUserCreateInput
