+++ src/interfaces/apollo/users/IUserUpdateInput.ts
@@ -2,8 +2,7 @@ interface IUserUpdateInput {
   _id: string
   email: string
   password: string
-  type: string
-  websiteIds: string[]
+  type: 'admin' | 'user'
 }
 
 export default IUserUpdateInput
