+++ src/interfaces/index.ts
@@ -1,30 +1,36 @@
-// general
+/* ----------------------------------------------------------------------------------
+ * Interfaces that is used anywhere in the app.
+/* ---------------------------------------------------------------------------------- */
 export { default as IDynamicObject } from './IDynamicObject'
 export { default as IError } from './IError'
 export { default as IExpressHook } from './IExpressHook'
 export { default as IReject } from './IReject'
 export { default as IResolve } from './IResolve'
 
-// apollo
-export { default as IAuthenticatedUser } from './apollo/IAuthenticatedUser'
-export { default as IIssue } from './apollo/IIssue'
+/* ----------------------------------------------------------------------------------
+ * Apollo Interfaces
+/* ---------------------------------------------------------------------------------- */
 export { default as IPaginatedData } from './apollo/IPaginatedData'
-export { default as IUserCreateInput } from './apollo/IUserCreateInput'
-export { default as IUserUpdateInput } from './apollo/IUserUpdateInput'
-export { default as IWebsite } from './apollo/IWebsite'
-export { default as IWebsiteInput } from './apollo/IWebsiteInput'
-export { default as IWebsitesPaginateInput } from './apollo/IWebsitesPaginateInput'
+// apollo > users
+export { default as IAuthenticatedUser } from './apollo/users/IAuthenticatedUser'
+export { default as IUserCreateInput } from './apollo/users/IUserCreateInput'
+export { default as IUserUpdateInput } from './apollo/users/IUserUpdateInput'
 
-// global
+/* ----------------------------------------------------------------------------------
+ * Global Directory Interfaces
+/* ---------------------------------------------------------------------------------- */
 export { default as IActiveRecordMatchValues } from './global/utilities/IActiveRecordMatchValues'
 export { default as IActiveRecordSearchFilter } from './global/utilities/IActiveRecordSearchFilter'
 
-// model
+/* ----------------------------------------------------------------------------------
+ * Mongoose Model Interfaces
+/* ---------------------------------------------------------------------------------- */
 export { default as IModelOptions } from './models/IModelOptions'
 export { default as IModelOptionsInput } from './models/IModelOptionsInput'
 export { default as ISingleDocModelOptions } from './models/ISingleDocModelOptions'
 export { default as ISingleDocModelOptionsInput } from './models/ISingleDocModelOptionsInput'
-export { default as IUser } from './models/IUser'
-export { default as IUserCreateInputPasswordHashed } from './models/IUserCreateInputPasswordHashed'
-export { default as IUserTypeOption } from './models/IUserTypeOption'
-export { default as IUserUpdateInputPasswordHashed } from './models/IUserUpdateInputPasswordHashed'
+// model > users
+export { default as IUser } from './models/users/IUser'
+export { default as IUserCreateInputPasswordHashed } from './models/users/IUserCreateInputPasswordHashed'
+export { default as IUserTypeOption } from './models/users/IUserTypeOption'
+export { default as IUserUpdateInputPasswordHashed } from './models/users/IUserUpdateInputPasswordHashed'
