+++ src/interfaces/models/users/IUserCreateInputPasswordHashed.ts
